Implement data bootstrapping harness
Before markup is rendered on the server, the necessary data must be loaded to
populate the react components. This needs to happen outside of react (which is
only a view layer, and should not handle state unrelated to views).

This creates a simple harness where data can be loaded (but does not implement
the loading or logic to select what to load yet). The data is passed to React
on the server, and then also included in the HTML so that it can be passed to
the client-side React as it boots up.