Add verification that token is valid
Instead of just checking that a token has been set in the cookies, this adds a
verification that the cookie is valid by making a simple request to the API for
the session object. If it fails, it removes the invalid token cookie and
redirects the visitor, now considered anonymous, to accounts.zetk.in to login.