Style close link for both layouts
In the horizontal layout, the close link is placed just over the pane, on top
of the header of the underlying pane. When the user taps the underlying pane,
they are actually hitting the close button of the overlying pane, resulting in
it in being closed which in turn focuses the pane the thought they tapped.