Remove campaign list sub-section
There should be no need for this now that we have the easy-to-access feature
for filtering campaign content by campaign. The entire idea of this section
was to access information about specific campaigns, but that can be done from
the overview instead, and campaign sub-content (such as actions) can now be
found by filtering in their respective areas instead.