Expand on dragOver, calculating height in JS
Previously a max-height hack was used in CSS to animate the expanding action
list items. This had many flaws, like having a fixed max height (set to 400px)
and animations not being a fixed duration. This update calculates the new
height in javascript and sets it as an inline style. It does so both when an
item is explicitly expanded and when it's hovered during a drag operation.