Rename widget data as "config" instead
The data that is given to the widget from the store is actually not the data
that it will be display, but rather it's configuration (currently no widget has
any configuration). The actual data will need to be retrieved by the widget,
and renaming the current "data" to config will reserve the "data" name for the
actual widget data.